+++ CHANGELOG.rst
@@ -11,6 +11,7 @@ Change Log
   - Added conflicts and recent histories to the top slideover of teams/adjudicators
   - Adding 'ranking' toggles to visibily highlight adjudicator strengths and more easily identify unbalanced panels
   - The adjudicator, venue, and matchup editing interfaces are now sortable
+- Re-implemented/added a chi-squared test to measure motion balance in the motion standings/balance
 - The Auto Allocate function for adjudicators now also allocates trainees to solo-chaired debates
 - Added a 'Tab Release' preset for easily releasing all standings/results pages after a tournament is finished
 - Fixed issue where toggling iron speeches on and off wouldn't hide/unset the relevant checkboxes